Move qpainter to a subpackage, add lots of QPainter methods
