+++ all.cpp
@@ -4,7 +4,6 @@
 #include "cpp/govaluetype.cpp"
 #include "cpp/idletimer.cpp"
 #include "cpp/connector.cpp"
-#include "cpp/painter.cpp"
 
 #include "cpp/moc_all.cpp"
