+++ qpainter/painter.cpp
@@ -1,119 +1,22 @@
-#ifndef PAINTER_H
-#define PAINTER_H
+// #include <private/qmetaobjectbuilder_p.h>
 
-// #include <private/qmetaobject_p.h>
+// #include <QtQml/QtQml>
+#include <QPainter>
+// #include <QQmlEngine>
+// #include <QDebug>
 
-// #include <QQuickPaintedItem>
-// #include <QPainter>
-
-#include "capi.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include "painter.h"
 
 
 
-// Enums stolen from <QPainter>
-// Cannot include <QPainter> since cgo always runs go stuff through gcc, not g++
-enum RenderHint {
-    Antialiasing = 0x01,
-    TextAntialiasing = 0x02,
-    SmoothPixmapTransform = 0x04,
-    HighQualityAntialiasing = 0x08,
-    NonCosmeticDefaultPen = 0x10,
-    Qt4CompatiblePainting = 0x20
-};
+void painterSetCompositionMode(QPainter_ *painter, CompositionMode mode) {
+  reinterpret_cast<QPainter *>(painter)->setCompositionMode((QPainter::CompositionMode) mode);
+}
+CompositionMode painterCompositionMode(QPainter_ *painter) {
+  return (CompositionMode) reinterpret_cast<QPainter *>(painter)->compositionMode();
+}
 
-enum CompositionMode {
-    CompositionMode_SourceOver,
-    CompositionMode_DestinationOver,
-    CompositionMode_Clear,
-    CompositionMode_Source,
-    CompositionMode_Destination,
-    CompositionMode_SourceIn,
-    CompositionMode_DestinationIn,
-    CompositionMode_SourceOut,
-    CompositionMode_DestinationOut,
-    CompositionMode_SourceAtop,
-    CompositionMode_DestinationAtop,
-    CompositionMode_Xor,
-
-    //svg 1.2 blend modes
-    CompositionMode_Plus,
-    CompositionMode_Multiply,
-    CompositionMode_Screen,
-    CompositionMode_Overlay,
-    CompositionMode_Darken,
-    CompositionMode_Lighten,
-    CompositionMode_ColorDodge,
-    CompositionMode_ColorBurn,
-    CompositionMode_HardLight,
-    CompositionMode_SoftLight,
-    CompositionMode_Difference,
-    CompositionMode_Exclusion,
-
-    // ROPs
-    RasterOp_SourceOrDestination,
-    RasterOp_SourceAndDestination,
-    RasterOp_SourceXorDestination,
-    RasterOp_NotSourceAndNotDestination,
-    RasterOp_NotSourceOrNotDestination,
-    RasterOp_NotSourceXorDestination,
-    RasterOp_NotSource,
-    RasterOp_NotSourceAndDestination,
-    RasterOp_SourceAndNotDestination,
-    RasterOp_NotSourceOrDestination,
-    RasterOp_SourceOrNotDestination,
-    RasterOp_ClearDestination,
-    RasterOp_SetDestination,
-    RasterOp_NotDestination
-};
-
-// Q_DECLARE_FLAGS(RenderHints, RenderHint)
-
-// class PixmapFragment {
-// public:
-//     qreal x;
-//     qreal y;
-//     qreal sourceLeft;
-//     qreal sourceTop;
-//     qreal width;
-//     qreal height;
-//     qreal scaleX;
-//     qreal scaleY;
-//     qreal rotation;
-//     qreal opacity;
-//     static PixmapFragment Q_GUI_EXPORT create(const QPointF &pos, const QRectF &sourceRect,
-//                                         qreal scaleX = 1, qreal scaleY = 1,
-//                                         qreal rotation = 0, qreal opacity = 1);
-// };
-
-// enum PixmapFragmentHint {
-//     OpaqueHint = 0x01
-// };
-
-// Q_DECLARE_FLAGS(PixmapFragmentHints, PixmapFragmentHint)
-
-// // QPainter();
-// // explicit QPainter(QPaintDevice *);
-// // ~QPainter();
 //
-// // QPaintDevice *painterDevice(QPainter_ *painter) const;
-//
-// // bool painterBegin(QPainter_ *painter, QPaintDevice *);
-// // bool painterEnd(QPainter_ *painter);
-// bool painterIsActive(QPainter_ *painter) const;
-//
-// void painterInitFrom(QPainter_ *painter, const QPaintDevice *device);
-//
-
-
-
-
-void painterSetCompositionMode(QPainter_ *painter, enum CompositionMode mode);
-enum CompositionMode painterCompositionMode(QPainter_ *painter);
-
 // const QFont &painterFont(QPainter_ *painter) const;
 // void painterSetFont(QPainter_ *painter, const QFont &f);
 //
@@ -161,8 +64,12 @@ enum CompositionMode painterCompositionMode(QPainter_ *painter);
 //
 // QRectF painterClipBoundingRect(QPainter_ *painter) const;
 //
-void painterSave(QPainter_ *painter);
-void painterRestore(QPainter_ *painter);
+void painterSave(QPainter_ *painter){
+  reinterpret_cast<QPainter *>(painter)->save();
+}
+void painterRestore(QPainter_ *painter){
+  reinterpret_cast<QPainter *>(painter)->restore();
+}
 //
 // // XForm functions
 // void painterSetMatrix(QPainter_ *painter, const QMatrix &matrix, bool combine = false);
@@ -190,21 +97,33 @@ void painterRestore(QPainter_ *painter);
 // void painterSetWorldMatrixEnabled(QPainter_ *painter, bool enabled);
 // bool painterWorldMatrixEnabled(QPainter_ *painter) const;
 //
-// void painterScale(QPainter_ *painter, qreal sx, qreal sy);
-// void painterShear(QPainter_ *painter, qreal sh, qreal sv);
-// void painterRotate(QPainter_ *painter, qreal a);
+void painterScale(QPainter_ *painter, qreal sx, qreal sy) {
+  reinterpret_cast<QPainter *>(painter)->scale(sx, sy);
+}
+void painterShear(QPainter_ *painter, qreal sh, qreal sv) {
+  reinterpret_cast<QPainter *>(painter)->shear(sh, sv);
+}
+void painterRotate(QPainter_ *painter, qreal a) {
+  reinterpret_cast<QPainter *>(painter)->rotate(a);
+}
 //
 // void painterTranslate(QPainter_ *painter, const QPointF &offset);
 // void painterTranslate(QPainter_ *painter, const QPoint &offset);
-// void painterTranslate(QPainter_ *painter, qreal dx, qreal dy);
+void painterTranslate(QPainter_ *painter, qreal dx, qreal dy) {
+  reinterpret_cast<QPainter *>(painter)->translate(dx, dy);
+}
 //
 // QRect painterWindow(QPainter_ *painter) const;
 // void painterSetWindow(QPainter_ *painter, const QRect &window);
-// void painterSetWindow(QPainter_ *painter, int x, int y, int w, int h);
+void painterSetWindow(QPainter_ *painter, int x, int y, int w, int h) {
+  reinterpret_cast<QPainter *>(painter)->setWindow(x, y, w, h);
+}
 //
 // QRect painterViewport(QPainter_ *painter) const;
 // void painterSetViewport(QPainter_ *painter, const QRect &viewport);
-// void painterSetViewport(QPainter_ *painter, int x, int y, int w, int h);
+void painterSetViewport(QPainter_ *painter, int x, int y, int w, int h) {
+  reinterpret_cast<QPainter *>(painter)->setViewport(x, y, w, h);
+}
 //
 // void painterSetViewTransformEnabled(QPainter_ *painter, bool enable);
 // bool painterViewTransformEnabled(QPainter_ *painter) const;
@@ -398,14 +317,11 @@ void painterRestore(QPainter_ *painter);
 // static QPaintDevice *redirected(const QPaintDevice *device, QPoint *offset = 0);
 // static void restoreRedirected(const QPaintDevice *device);
 
-void painterBeginNativePainting(QPainter_ *painter);
-void painterEndNativePainting(QPainter_ *painter);
-
-
-#ifdef __cplusplus
-} // extern "C"
-#endif
-
-#endif // GOVALUE_H
+void painterBeginNativePainting(QPainter_ *painter) {
+  reinterpret_cast<QPainter *>(painter)->beginNativePainting();
+}
+void painterEndNativePainting(QPainter_ *painter) {
+  reinterpret_cast<QPainter *>(painter)->endNativePainting();
+}
 
 // vim:ts=4:sw=4:et:ft=cpp
